requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
CurrencyConverter==0.16.11
discord.py==1.7.3
elastic-transport==8.1.2
elasticsearch==8.1.3
elasticsearch-async==6.2.0
idna==3.3
multidict==6.0.2
requests==2.27.1
soupsieve==2.3.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
